# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "AgentToll",
    "version": "0.15.0",
    "description": "Base mainnet data with x402 payments. Hosted prices: $0.001-$0.008 USDC. Self-hosted payment network and recipient are configured at startup; inspect /api/catalog and /.well-known/agent-card.json. PAYMENT-SIGNATURE sends authorization; PAYMENT-REQUIRED and PAYMENT-RESPONSE carry the quote and receipt.",
    "contact": {
      "name": "AgentToll",
      "url": "https://github.com/tevfikefeaydin/agenttoll/issues",
      "email": "tevfikefe.aydin@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://github.com/tevfikefeaydin/agenttoll/blob/main/LICENSE"
    }
  },
  "servers": [
    {
      "url": "https://agenttoll.app"
    }
  ],
  "paths": {
...
    "/api/price/{symbol}": {
      "get": {
        "summary": "Token price lookup: spot price in USD and 24h change for a crypto asset, by ticker or CoinGecko id",
        "x-payment-info": {
          "price": {
            "currency": "USD",
            "mode": "fixed",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "network": "eip155:8453",
                "scheme": "exact",
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "assetSymbol": "USDC",
                "payTo": "0xe55359021a6a22d8385b827405991c56075f56f8"
              }
            }
          ]
        },
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Ticker or CoinGecko id"
            },
            "example": "eth"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. meta describes response freshness and configured networks.",
            "content": {
              "application/json": {
                "example": {
                  "symbol": "eth",
                  "id": "ethereum",
                  "usd": 1867.28,
                  "change24h": -0.1,
                  "source": "coingecko",
                  "at": "2026-08-05T13:31:57.070Z",
                  "quoteCurrency": "USD",
                  "assumptions": [],
                  "meta": {
                    "requestId": "example-request",
                    "servedAt": "2026-08-05T13:31:57.070Z",
                    "observedAt": "2026-08-05T13:31:57.070Z",
                    "ageSeconds": 0,
                    "dataNetwork": "base",
                    "paymentNetwork": "base"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "402": {
            "description": "Unsigned x402 v2 quote in the base64 JSON PAYMENT-REQUIRED header.",
            "headers": {
              "PAYMENT-REQUIRED": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Upstream unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Request deadline exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "operationId": "get_price"
      }
    },
    "/api/gas": {
      "get": {
        "summary": "Base gas price check: the current gas price and latest block number; add ?gasLimit=150000 to also get what a transaction that size costs in ETH and USD",
        "x-payment-info": {
          "price": {
            "currency": "USD",
            "mode": "fixed",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "network": "eip155:8453",
                "scheme": "exact",
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "assetSymbol": "USDC",
                "payTo": "0xe55359021a6a22d8385b827405991c56075f56f8"
              }
            }
          ]
        },
        "parameters": [
          {
            "name": "gasLimit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 21000,
              "maximum": 30000000,
              "description": "Optional. Also price a transaction of this gas size"
            },
            "example": 150000
          }
        ],
        "responses": {
          "200": {
            "description": "Success. meta describes response freshness and configured networks.",
            "content": {
              "application/json": {
                "example": {
                  "chain": "base",
                  "gasPriceWei": "6041753",
                  "gasPriceGwei": 0.006041753,
                  "latestBlock": 49573693,
                  "estimate": {
                    "gasLimit": 150000,
                    "ethCost": 9.06263e-7,
                    "usdCost": 0.001692,
                    "ethUsd": 1867.28
                  },
                  "at": "2026-08-05T13:32:13.596Z",
                  "meta": {
                    "requestId": "example-request",
                    "servedAt": "2026-08-05T13:32:13.596Z",
                    "observedAt": "2026-08-05T13:32:13.596Z",
                    "ageSeconds": 0,
                    "dataNetwork": "base",
                    "paymentNetwork": "base"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "402": {
            "description": "Unsigned x402 v2 quote in the base64 JSON PAYMENT-REQUIRED header.",
            "headers": {
              "PAYMENT-REQUIRED": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Upstream unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Request deadline exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "operationId": "get_base_gas"
      }
    },
    "/api/trending": {
      "get": {
        "summary": "Trending token scan: the tokens moving across the crypto market right now; ?limit=N trims the list",
        "x-payment-info": {
          "price": {
            "currency": "USD",
            "mode": "fixed",
            "amount": "0.002"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "network": "eip155:8453",
                "scheme": "exact",
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "assetSymbol": "USDC",
                "payTo": "0xe55359021a6a22d8385b827405991c56075f56f8"
              }
            }
          ]
        },
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 25,
              "description": "Optional. Top N only"
            },
            "example": 5
          }
        ],
        "responses": {
          "200": {
            "description": "Success. meta describes response freshness and configured networks.",
            "content": {
              "application/json": {
                "example": {
                  "coins": [
                    {
                      "id": "hyperliquid",
                      "symbol": "hype",
                      "name": "Hyperliquid",
                      "rank": 18,
                      "usd": 22.4,
                      "change24h": 3.1
                    }
                  ],
                  "source": "coingecko",
                  "at": "2026-08-05T13:31:58.678Z",
                  "meta": {
                    "requestId": "example-request",
                    "servedAt": "2026-08-05T13:31:58.678Z",
                    "observedAt": "2026-08-05T13:31:58.678Z",
                    "ageSeconds": 0,
                    "dataNetwork": "base",
                    "paymentNetwork": "base"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "402": {
            "description": "Unsigned x402 v2 quote in the base64 JSON PAYMENT-REQUIRED header.",
            "headers": {
              "PAYMENT-REQUIRED": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Upstream unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Request deadline exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "operationId": "get_trending"
      }
    },
    "/api/base/token/{address}": {
      "get": {
        "summary": "Base token price lookup by contract address: the onchain USD price for any token on Base",
        "x-payment-info": {
          "price": {
            "currency": "USD",
            "mode": "fixed",
            "amount": "0.001"
          },
          "protocols": [
            {
              "x402": {
                "version": 2,
                "network": "eip155:8453",
                "scheme": "exact",
                "asset": "0x833589fCD6eDb6E08f4c7C32D4f71b54bdA02913",
                "assetSymbol": "USDC",
                "payTo": "0xe55359021a6a22d8385b827405991c56075f56f8"
              }
            }
          ]
        },
        "parameters": [
          {
            "name": "address",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^0x[0-9a-fA-F]{40}$"
            },
            "example": "0x4200000000000000000000000000000000000006"
          }
        ],
        "responses": {
          "200": {
            "description": "Success. meta describes response freshness and configured networks.",
            "content": {
              "application/json": {
                "example": {
                  "chain": "base",
                  "token": "0x4200000000000000000000000000000000000006",
                  "usd": 1867.28,
                  "source": "geckoterminal",
                  "at": "2026-08-05T13:31:58.678Z",
                  "meta": {
                    "requestId": "example-request",
                    "servedAt": "2026-08-05T13:31:58.678Z",
                    "observedAt": "2026-08-05T13:31:58.678Z",
                    "ageSeconds": 0,
                    "dataNetwork": "base",
                    "paymentNetwork": "base"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "402": {
            "description": "Unsigned x402 v2 quote in the base64 JSON PAYMENT-REQUIRED header.",
            "headers": {
              "PAYMENT-REQUIRED": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "429": {
            "description": "Rate limited",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "502": {
            "description": "Upstream unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "504": {
            "description": "Request deadline exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "operationId": "get_base_token_price"
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "required": [
          "error",
          "code",
          "retryable",
          "retryAfter",
          "requestId"
        ],
        "properties": {
          "error": {
            "type": "string"
          },
          "code": {
            "type": "string"
          },
          "retryable": {
            "type": "boolean"
          },
          "retryAfter": {
            "type": "integer",
            "nullable": true
          },
          "requestId": {
            "type": "string"
          }
        }
      }
    }
  }
}